Add files via upload
The project is the Business Understanding which is developed by carefully studying the operations of At Home Group as stated on their 2019 Financial Results Report. From this analysis it is possible to determine what is it that the company needs, which in the present case it was believed to be a method to determine or estimate if a store will have successful results.
Following the previous phase, the next step of the project covers the Data Gathering and Preparation. To do so, the data points extracted are related to demographic and socio-economic factors of the zip codes in which the stores are located from several governmental sources and related sources. Then, there is a phase of Exploratory Data Analysis which stimulates the development of basic data understanding, followed by a section of Feature Engineering to maximize the potential usability of the available features.
Last, there is the Modeling section in which is possible to find a supervised machine learning model that helps satisfy the business need declared above, as well as the Evaluation of it. It is important to clarify that the last phase of the CRISP-DM project cycle includes the Deployment stage.
Once we have performed all the calculations and modeling throughout the project, we can conclude with the following ideas. First, we can state that the model has been able to predict the average sales growth of At Home Group stores. The support vector regressor had the best performance metrics between the three models created. This model was able to calculate the average sales growth value by store with an error or an average difference of only 0.20 rate value. In other words, the difference between the real value and the predicted one for the average sales growth will be of 20%.